Module webio::callback::multi

This module implements conversion from callbacks to futures for callbacks that are called multiple times, i.e. callbacks of events that occur more than once per callback.

Structs

AsyncRegister

Register of multi-call callbacks into an event, where the callback is asynchronous (waiting for the callback to complete is also asynchronous).

ListenNext

A handle to wait for the single next occurence of an event and a registered callback.

Listener

A handle to a multi-call callback registered in an event. Typically, the Listener is used with the Listener::listen_next method for awaiting next occurences of an event, but it can also be used as a stream.

SyncRegister

Register of multi-call callbacks into an event, where the callback is synchronous (though waiting for the callback to complete is still asynchronous).

Type Definitions

AsyncCbHandler

The type of asynchronous, multi-call callback handlers (i.e. the handler that calls callbacks): a boxed mutable function yielding boxed futures, a wrapper over callbacks.

AsyncCbHandlerFuture

The type of futures used in asynchronous, multi-call callback handlers (i.e. the handler that calls callbacks): a boxed future.

SyncCbHandler

The type of synchronous, multi-call callback handlers (i.e. the handler that calls callbacks): a boxed mutable function, a wrapper over callbacks.