Derive Macro webio::EventType

#[derive(EventType)]
{
    // Attributes available to this derive:
    #[event_type]
}

Available on crate features macros and event only.

Defines a custom event wrapper, with the intention of being safe. It is up to the caller type, however, to ensure that name is correct for the given event data type.

It is required that event_type(name = ..., data = ...) attribute is required, and should be placed at the top of the struct; it is ignored elsewhere.

Example

#[derive(EventType)]
#[event_type(name = "click", data = web_sys::MouseEvent)]
struct CustomClick;

let listener = CustomClick.add_listener(&element);

element.dispatch_event(&web_sys::MouseEvent::new("click").unwrap()).unwrap();
listener.listen_next().await.unwrap();
element.dispatch_event(&web_sys::MouseEvent::new("click").unwrap()).unwrap();
listener.listen_next().await.unwrap();