Function webio::event::add_listener_raw

pub fn add_listener_raw<S, E>(
    target: &EventTarget, 
    event_type: S
) -> Listener<E> where
    S: Into<String>,
    E: FromWasmAbi + 'static, 

Available on crate feature event only.

Raw function for adding event listeners to JS’s event targets. This function is asynchronous and a future is returned.

It is up to the caller to ensure that the event_type is correct and generic parameter E matches the event_type, as well to ensure the target supports such event_type.