Function webio::event::add_sync_listener_raw

pub fn add_sync_listener_raw<F, E, T>(
    target: &EventTarget, 
    event_type: &str, 
    callback: F
) -> Listener<T> where
    F: FnMut(E) -> T + 'static,
    E: FromWasmAbi + 'static,
    T: 'static, 

Available on crate feature event only.

Raw function for adding event listeners to JS’s event targets, using synchronous event listeners. However, this function is asynchronous and a future is returned.

It is up to the caller to ensure that the event_type is correct and generic parameter E matches the event_type, as well to ensure the target supports such event_type.