Trait webio::event::EventType

pub trait EventType {
    type Data: FromWasmAbi + 'static;

    fn name(&self) -> String;

    fn add_listener(&self, target: &EventTarget) -> Listener<Self::Data> { ... }
    fn add_listener_with_sync_cb<F, T>(
        &self, 
        target: &EventTarget, 
        callback: F
    ) -> Listener<T>
    where
        F: FnMut(Self::Data) -> T + 'static,
        T: 'static,
    { ... }
    fn add_listener_with_async_cb<F, A>(
        &self, 
        target: &EventTarget, 
        callback: F
    ) -> Listener<A::Output>
    where
        F: FnMut(Self::Data) -> A + 'static,
        A: Future + 'static,
    { ... }
}

Available on crate feature event only.

Trait for safe wrappers over JS event types and JS event listening.

It is up to the implementor to ensure that the event_type is correct and associated type Data matches .name().

Required Associated Types

type Data: FromWasmAbi + 'static

Data of an event’s occurence, passed to the listener.

Required Methods

fn name(&self) -> String

Name of this event type.

Provided Methods

fn add_listener(&self, target: &EventTarget) -> Listener<Self::Data>

Adds event listeners to JS’s event targets, where events are of this event type. This function is asynchronous and a future is returned.

It is up to the caller to ensure the target supports this event type.

fn add_listener_with_sync_cb<F, T>(
&self,
target: &EventTarget,
callback: F
) -> Listener<T> where
F: FnMut(Self::Data) -> T + 'static,
T: 'static,

Adds event listeners to JS’s event targets, where events are of this event type, using synchronous event listeners. However, this function is asynchronous and a future is returned.

It is up to the caller to ensure the target supports this event type.

fn add_listener_with_async_cb<F, A>(
&self,
target: &EventTarget,
callback: F
) -> Listener<A::Output> where
F: FnMut(Self::Data) -> A + 'static,
A: Future + 'static,

Adds event listeners to JS’s event targets, where events are of this event type, using asynchronous event listeners. This function is asynchronous and a future is returned.

It is up to the caller to ensure the target supports this event type.