Trait webio::event::Type

pub trait Type {
    type Data: FromWasmAbi + 'static;
    fn name(&self) -> &str;

    fn add_sync_listener<F, T>(
        &self, 
        target: &EventTarget, 
        callback: F
    ) -> Listener<T>
    where
        F: FnMut(Self::Data) -> T + 'static,
        T: 'static,
    { ... }

    fn add_async_listener<F, A>(
        &self, 
        target: &EventTarget, 
        callback: F
    ) -> Listener<A::Output>
    where
        F: FnMut(Self::Data) -> A + 'static,
        A: Future + 'static,
    { ... }
}

Trait for safe wrappers over JS event types and JS event listening.

It is up to the implementor to ensure that the event_type is correct and associated type Data matches .name().

Associated Types

type Data: FromWasmAbi + 'static

Data of an event’s occurence, passed to the listener.

Required methods

fn name(&self) -> &str

Name of this event type.

Provided methods

fn add_sync_listener<F, T>(
&self,
target: &EventTarget,
callback: F
) -> Listener<T> where
F: FnMut(Self::Data) -> T + 'static,
T: 'static,

Adds event listeners to JS’s event targets, where events are of this event type, using synchronous event listeners. However, this function is asynchronous and a future is returned.

It is up to the caller to ensure the target supports this event type.

fn add_async_listener<F, A>(
&self,
target: &EventTarget,
callback: F
) -> Listener<A::Output> where
F: FnMut(Self::Data) -> A + 'static,
A: Future + 'static,

Adds event listeners to JS’s event targets, where events are of this event type, using asynchronous event listeners. This function is asynchronous and a future is returned.

It is up to the caller to ensure the target supports this event type.