Macro webio::join

join!() { /* proc-macro */ }

Available on crate feature macros only.

Joins a list of futures and returns their output into a tuple in the same order that the futures were given. Futures must be 'static.

Syntax:

join!(future0, future1, future2, ..., future_n)

Examples

With Timeout

use std::time::Duration;
use webio::{join, time::timeout};
  

// Create some tasks
let first_handle = async {
    timeout(Duration::from_millis(50)).await;
    3
};
let second_handle = async {
    timeout(Duration::from_millis(60)).await;
    5
};
let third_handle = async {
    timeout(Duration::from_millis(40)).await;
    7
};

// Join them
let (first, second, third) = join!(first_handle, second_handle, third_handle);

// Expected output
assert_eq!((first, second, third), (3, 5, 7));