Macro webio::select

select!() { /* proc-macro */ }

Available on crate feature macros only.

Listens to a list of futures and finishes when the first future finishes, which is then selected. Every future is placed in a “match arm”, and when it is selected, the “arm” pattern is matched and the macro evaluates to the right side of the “arm”. Patterns must be irrefutable, typically just a variable name, or destructuring. Futures must be 'static.

Syntax:

select! {
    pattern0 = future0 => output0,
    pattern1 = future1 => output1,
    pattern2 = future2 => output2,
    ...,
    pattern_n = future_n => output_n,
}

Examples

With Timeout

use std::time::Duration;
use webio::{select, time::timeout};

// Create some tasks
let first_handle = async {
    timeout(Duration::from_millis(500)).await;
    3u32
};
let second_handle = async {
    timeout(Duration::from_millis(50)).await;
    5u32
};
let third_handle = async {
    timeout(Duration::from_millis(350)).await;
    7u32
};

// Select the first one to complete
let output = select! {
    val = first_handle => val + 10,
    val = second_handle => val + 20,
    val = third_handle => val - 5
};

// Second one should be the first
assert_eq!(output, 25);