Macro webio::try_join

try_join!() { /* proc-macro */ }

Available on crate feature macros only.

Joins a list of futures and returns their output into a tuple in the same order that the futures were given, but if one of them fails, try_join fails, and so a result of tuples is returned. Futures must be 'static.

Syntax:

try_join!(future0, future1, future2, ..., future_n)

Examples

With Timeout and Success

use std::time::Duration;
use webio::{try_join, time::timeout};

// Create some tasks
let first_handle = async {
    timeout(Duration::from_millis(50)).await;
    Result::<u32, &str>::Ok(3)
};
let second_handle = async {
    timeout(Duration::from_millis(60)).await;
    Ok(5)
};
let third_handle = async {
    timeout(Duration::from_millis(40)).await;
    Ok(7)
};

// Try to join it
let result = try_join!(first_handle, second_handle, third_handle);
// Should be Ok
let (first, second, third) = result.unwrap();

// Expected output
assert_eq!((first, second, third), (3, 5, 7));

With Timeout and Failure

use std::time::Duration;
use webio::{try_join, time::timeout};

// Create some tasks
let first_handle = async {
    timeout(Duration::from_millis(50)).await;
    Ok(3)
};
let second_handle = async {
    timeout(Duration::from_millis(60)).await;
    Err("boom")
};
let third_handle = async {
    timeout(Duration::from_millis(40)).await;
    Ok(7)
};

// Try to join them
let result = try_join!(first_handle, second_handle, third_handle);

// Should be an error
assert_eq!(result, Err("boom"));