Module webio::sync

This module provides locks to be used within a single instance of a Rust WebAssembly module. Why locks given that a WASM module is single-threaded? Well, when using async Rust, a critical operation can be split by an .await expression. In fact, these locks are designed specially for that: ensuring a critical operation is performed as if it were atomic even if you insert an .await between two steps.

Structs

Mutex

Mutual exclusion of critical sections. Behaves much like [tokio::sync::Mutex], but designed for WASM (single-thread, thus this struct is Unsync). This mutex is fair.

MutexGuard

A guard of a current locking on a Mutex. Can be derreferenced to get access to protected data.

ReadGuard

A guard of a current read/shared-locking on a RwLock. Can be derreferenced to get read access to protected data.

RwLock

Read-Write lock over critical sections where data can be locked shared or exclusively. When locked shared, only reads are allowed, while when locked exclusively, writes are also allowed. Behaves much like [tokio::sync::RwLock], but designed for WASM (single-thread, thus this struct is Unsync). This lock is fair.

WriteGuard

A guard of a current write/exclusive-locking on a RwLock. Can be derreferenced to get both read and write access to protected data.