Module webio::task

This module exports items related to task spawning.

Structs

JoinError

An error that might happen when waiting for a task.

JoinHandle

A handle that allows the caller to join a task (i.e. wait for it to end).

Functions

detach

Detaches a future from the current WASM call, but ensures the future completes.

spawn

Spawns an asynchronous task in JS event loop.

yield_now

Yields control back to the event loop once and returns back to execution as soon as possible.