Module webio::time

This module implements time-related utilities.

Structs

Instant

A montonic clock measurement, mimicking std::time::Instant but for WASM, which is not supported by std’s Instant. Behind the curtains, this type uses JavaScript’s performance::now(), so there is an overhead when calling Instant::now.

TimeoutHandle

A handle to a timeout call. The timeout can be waited through .await, or it can be cancelled when the handle is dropped without the timeout completing.

Functions

timeout

Creates a Future that completes only after some duration of time has passed.