Trait webio::event::Type

pub trait Type {
    type Data: FromWasmAbi + 'static;

    fn name(&self) -> &str;

    fn add_listener(&self, target: &EventTarget) -> Listener<Self::Data> { ... }
}

Available on crate feature event only.

Trait for safe wrappers over JS event types and JS event listening.

It is up to the implementor to ensure that the event_type is correct and associated type Data matches .name().

Required Associated Types

type Data: FromWasmAbi + 'static

Data of an event’s occurence, passed to the listener.

Required Methods

fn name(&self) -> &str

Name of this event type.

Provided Methods

fn add_listener(&self, target: &EventTarget) -> Listener<Self::Data>

Adds event listeners to JS’s event targets, where events are of this event type. This function is asynchronous and a future is returned.

It is up to the caller to ensure the target supports this event type.